Commit change resources

+++ services/web/project/resources/directors.py
@@ -3,6 +3,7 @@ Director entity resources
 """
 from flask import request
 from flask_restx import Resource
+from flask_login import login_required, current_user
 
 from ..modelschemas.directors import DirectorSchema
 from ..models.directors import Director
@@ -17,7 +18,7 @@ class DirectorList(Resource):
     """Director Data Resource Class"""
 
     @classmethod
-    def get(cls):
+    def get(cls) -> (dict, int):
         """
         Get method
         :return: error message or json with information about the directors
@@ -25,23 +26,26 @@ class DirectorList(Resource):
         return director_list_schema.dump(Director.find_all()), 200
 
     @classmethod
-    def post(cls):
+    @login_required
+    def post(cls) -> (dict, int):
         """
         Post method
         :return: json with information about the directors
         """
-        director_json = request.get_json()
-        director_data = director_schema.load(director_json, session=db.session)
-        director_data.save_to_db()
+        if current_user.is_authenticated and current_user.is_admin:
+            director_json = request.get_json()
+            director_data = director_schema.load(director_json, session=db.session)
+            director_data.save_to_db()
 
-        return director_schema.dump(director_data), 201
+            return director_schema.dump(director_data), 201
+        return {"status": 401, "reason": "User is not admin"}
 
 
 class DirectorListId(Resource):
     """Director Data Resource Class"""
 
     @classmethod
-    def get(cls, director_id):
+    def get(cls, director_id) -> (dict, int):
         """
         Get method
         :param director_id: director`s id
@@ -50,36 +54,42 @@ class DirectorListId(Resource):
         director_data = Director.query.get_or_404(director_id)
         if director_data:
             return director_schema.dump(director_data), 200
-        return {'message': DIRECTOR_NOT_FOUND}, 404
+        return {"status": 404, 'message': DIRECTOR_NOT_FOUND}
 
     @classmethod
-    def delete(cls, director_id):
+    @login_required
+    def delete(cls, director_id) -> dict:
         """
         Delete method
         :param director_id: director`s id
         :return: error message or successful message
         """
-        director_data = Director.query.get_or_404(director_id)
-        if director_data:
-            director_data.delete_from_db()
-            return {'message': "Director Deleted successfully"}, 200
-        return {'message': DIRECTOR_NOT_FOUND}, 404
+        if current_user.is_authenticated and current_user.is_admin:
+            director_data = Director.query.get_or_404(director_id)
+            if director_data:
+                director_data.delete_from_db()
+                return {"status": 200, 'message': "Director Deleted successfully"}
+            return {"status": 404, 'message': DIRECTOR_NOT_FOUND}
+        return {"status": 401, "reason": "User is not admin"}
 
     @classmethod
-    def put(cls, director_id):
+    @login_required
+    def put(cls, director_id) -> (dict, int):
         """
         Put method
         :param director_id: director`s id
         :return: json with information about the directors
         """
-        director_data = Director.query.get_or_404(director_id)
-        director_json = request.get_json()
+        if current_user.is_authenticated and current_user.is_admin:
+            director_data = Director.query.get_or_404(director_id)
+            director_json = request.get_json()
 
-        if director_data:
-            director_data.dir_first_name = director_json['dir_first_name']
-            director_data.dir_last_name = director_json['dir_last_name']
-        else:
-            director_data = director_schema.load(director_json, session=db.session)
+            if director_data:
+                director_data.dir_first_name = director_json['dir_first_name']
+                director_data.dir_last_name = director_json['dir_last_name']
+            else:
+                director_data = director_schema.load(director_json, session=db.session)
 
-        director_data.save_to_db()
-        return director_schema.dump(director_data), 200
+            director_data.save_to_db()
+            return director_schema.dump(director_data), 200
+        return {"status": 401, "reason": "User is not admin"}
