+++ services/web/project/resources/films.py
@@ -3,6 +3,7 @@ Director entity resources
 """
 from flask import request
 from flask_restx import Resource
+from flask_login import login_required, current_user
 
 from ..modelschemas.films import FilmSchema
 from ..models.films import Film
@@ -18,7 +19,7 @@ class FilmList(Resource):
     """Film Data Resource Class"""
 
     @classmethod
-    def get(cls):
+    def get(cls) -> (dict, int):
         """
         Get method
         :return: error message or json with information about the film
@@ -26,23 +27,26 @@ class FilmList(Resource):
         return film_list_schema.dump(Film.find_all()), 200
 
     @classmethod
-    def post(cls):
+    @login_required
+    def post(cls) -> (dict, int):
         """
         Post method
         :return: json with information about the film
         """
-        film_json = request.get_json()
-        film_data = film_schema.load(film_json, session=db.session)
-        film_data.save_to_db()
+        if current_user.is_authenticated and current_user.is_admin:
+            film_json = request.get_json()
+            film_data = film_schema.load(film_json, session=db.session)
+            film_data.save_to_db()
 
-        return film_schema.dump(film_data), 201
+            return film_schema.dump(film_data), 201
+        return {"status": 401, "reason": "User is not admin"}
 
 
 class FilmListId(Resource):
     """Film Data Resource Class"""
 
     @classmethod
-    def get(cls, film_id):
+    def get(cls, film_id) -> (dict, int):
         """
         Get method
         :param film_id: film`s id
@@ -51,10 +55,11 @@ class FilmListId(Resource):
         film_data = Film.query.get_or_404(film_id)
         if film_data:
             return film_schema.dump(film_data), 200
-        return {'message': FILM_NOT_FOUND}, 404
+        return {"status": 404, 'message': FILM_NOT_FOUND}
 
     @classmethod
-    def delete(cls, film_id):
+    @login_required
+    def delete(cls, film_id) -> dict:
         """
         Delete method
         :param film_id: film`s id
@@ -62,12 +67,17 @@ class FilmListId(Resource):
         """
         film_data = Film.query.get_or_404(film_id)
         if film_data:
-            film_data.delete_from_db()
-            return {'message': "Film Deleted successfully"}, 200
-        return {'message': FILM_NOT_FOUND}, 404
+            if current_user.is_authenticated:
+                if current_user.is_admin or current_user.user_id == film_data.user_id:
+                    film_data.delete_from_db()
+                    return {"status": 200, 'message': "Film Deleted successfully"}
+                return {"status": 401, "reason": "User is not admin or film owner"}
+            return {"status": 401, "reason": "User is not authenticated"}
+        return {"status": 404, 'message': FILM_NOT_FOUND}
 
     @classmethod
-    def put(cls, film_id):
+    @login_required
+    def put(cls, film_id) -> (dict, int):
         """
         Put method
         :param film_id: film`s id
@@ -77,12 +87,17 @@ class FilmListId(Resource):
         film_json = request.get_json()
 
         if film_data:
-            film_data.user_id = film_json['user_id']
-            film_data.film_name = film_json['film_name']
-            film_data.release_date = film_json['release_date']
-            film_data.description = film_json['description']
-            film_data.rating = film_json['rating']
-            film_data.poster = film_json['poster']
+            if current_user.is_authenticated:
+                if current_user.is_admin or current_user.user_id == film_data.user_id:
+                    film_data.film_name = film_json['film_name']
+                    film_data.release_date = film_json['release_date']
+                    film_data.description = film_json['description']
+                    film_data.rating = film_json['rating']
+                    film_data.poster = film_json['poster']
+                else:
+                    return {"status": 401, "reason": "User is not admin or film owner"}
+            else:
+                return {"status": 401, "reason": "User is not authenticated"}
         else:
             film_data = film_schema.load(film_json, session=db.session)
